refactor(app): curve out View/Model to node modules
