+++ public/index.html
@@ -5,7 +5,7 @@
   <link rel="stylesheet" type="text/css" href="css/bootstrap.css">
   <script src="js/jquery.js"></script>
   <script src="js/backbone.js"></script>
-  <script src="js/app.js"></script>
+  <script src="js/build/app.js"></script>
 </head>
 <body>
 <div class="container">
